Remove mcd transformer & shared repo/converter
