+++ transformers/test_data/tend.go
@@ -79,5 +79,5 @@ var tendModel = event.InsertionModel{
 }
 
 func TendModel() event.InsertionModel {
-	return CopyEventModel(tendModel)
+	return CopyModel(tendModel)
 }
